\name{cMonkeyNwInf-package}
\alias{cMonkeyNwInf-package}
\alias{cMonkeyNwInf}
\docType{package}
\title{
Inferelator-like network inference on cMonkey biclusters
}
\description{
Inferelator-like network inference on cMonkey biclusters
}
\details{
\tabular{ll}{
Package: \tab cMonkeyNwInf\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-08\cr
License: \tab GPL-2\cr
}
\url{}
}
\author{
David J Reiss, Institute for Systems Biology

Maintainer: \email{dreiss@systemsbiology.org}
}
\references{
 
}
\keyword{ package }
\seealso{
\url{}
}
\examples{
 
}
